updatetje

+++ powercalculations/_powerflows.py
@@ -1,4 +1,4 @@
-def power_flow(self, max_charge: float = 8, max_AC_power_output: float = 2, max_DC_batterypower: float = 2, max_PV_input: float = 10):
+def power_flow(self, max_charge: int = 8, max_AC_power_output: int = 2, max_DC_batterypower: int = 2, max_PV_input: int = 10):
     """
     Calculates power flows, how much is going to and from the battery and how much is being tapped from the grid
     #TODO: add units, PV_generated_power and Load_kW are both in kW. Depending on the frequency of this data, a different amount is subtracted from the battery charge (in kWh?) (e.g. if 1h freq, the load of each line can be subtracted directly since 1kW*1h=1kWh. If in minutes, then 1kW*1min=1/60kWh)
